requirements.txt
absl-py==2.0.0
appdirs==1.4.4
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.1
gitdb==4.0.11
GitPython==3.1.40
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
icecream==2.1.3
idna==3.6
imageio==2.33.0
ipython==8.18.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
line-profiler==4.1.2
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
parso==0.8.3
pettingzoo==1.24.2
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyastar2d==1.0.6
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
seaborn==0.13.0
sentry-sdk==1.38.0
setproctitle==1.3.3
six==1.16.0
slackweb==1.0.5
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tensorboardX==2.6.2.2
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.14.0
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.1
wcwidth==0.2.12
widgetsnbextension==4.0.9
